requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
idna==3.6
joblib==1.3.2
kafka-python==2.0.2
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
praw==7.7.1
prawcore==2.4.0
py4j==0.10.9.7
pymongo==4.6.2
pyspark==3.5.1
python-dateutil==2.8.2
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
textblob==0.18.0.post0
tqdm==4.66.2
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.7.0
